pipe/pyproject.toml
[project]
name = "trading-signals-pipeline"
version = "0.1.0"
description = "Prefect pipeline for automated trading signal generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Prefect Workflow Orchestration
    "prefect>=2.19.0",
    "prefect-github>=0.3.1",
    # Data Fetching
    "requests>=2.31.0",
    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.2",
    # Database (Python 3.13 compatible)
    "sqlalchemy>=2.0.35",
    "psycopg[binary,pool]>=3.1.0",
    # Email Service
    "resend==0.7.0",
    # Utilities
    "python-dotenv==1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = []
